refactoring of BasicChart.jsx done
