+++ src/VizG.jsx
@@ -26,7 +26,6 @@ import TableCharts from './components/TableChart.jsx';
 import NumberCharts from './components/NumberChart.jsx';
 import InlineCharts from './components/InlineChart.jsx';
 import VizGError from './VizGError';
-import BR from './components/BasicChartRefactor';
 
 class VizG extends Component {
 
@@ -84,7 +83,7 @@ class VizG extends Component {
             case 'area':
             case 'bar':
                 return (
-                    <BR
+                    <BasicCharts
                         config={config} metadata={metadata}
                         data={data} onClick={onClick}
                         yDomain={this.props.yDomain}
